test: fix unexpected token in BaseLayerRightToolbar.test.ts.
